requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-filters==1.0.0.dev0
djangorestframework-simplejwt==4.6.0
idna==2.10
oauthlib==3.1.0
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
